backend/requirements.txt
# Framework e Servidor
fastapi==0.111.1
uvicorn==0.30.1
gunicorn==22.0.0
starlette>=0.37.2,<0.38.0

# Google Cloud & Firebase
firebase-admin==6.5.0
google-cloud-firestore==2.16.0

# Autenticação e Segurança
python-dotenv>=1.0.1
pyjwt>=2.8.0
passlib>=1.7.4
python-jose>=3.3.0

# Pagamentos
stripe==10.9.0
paypalrestsdk>=1.13.3

# Utilitários e Validação
pydantic>=2.6.4
email-validator>=2.2.0
requests>=2.31.0
python-multipart>=0.0.9
Pillow>=10.4.0

# Google APIs (se usar Calendar, etc.)
google-api-python-client>=2.108.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0